# ARROW-11482: [Rust] improve take kernel performance of PrimitiveArray<T>