This PR simplifies the `take_primitive` a little bit and increases performance by using the new `from_trusted_len_iterator` API.

```
take i32 512            time:   [678.68 ns 692.27 ns 709.58 ns]                          
                        change: [-11.636% -10.615% -9.2359%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  5 (5.00%) high mild
  10 (10.00%) high severe

take i32 1024           time:   [1.1393 us 1.1477 us 1.1637 us]                           
                        change: [-18.397% -17.829% -17.075%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  5 (5.00%) high severe

take i32 nulls 512      time:   [643.87 ns 644.11 ns 644.37 ns]                                
                        change: [-20.604% -20.531% -20.458%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) low mild
  6 (6.00%) high mild
  5 (5.00%) high severe

take i32 nulls 1024     time:   [1.1381 us 1.1384 us 1.1387 us]                                 
                        change: [-19.419% -19.054% -18.753%] (p = 0.00 < 0.05)
                        Performance has improved.
```