@ritchie46 2021-02-03
> Thanks @ritchie46 !
> 
> Note that this collides with #9301 .

Right, great work on that PR! :) Then I will close this one.

I am also working on a perf increase of take_string. That one doesn't have collisions I presume?