@jorgecarleitao 2021-02-03
> Right, great work on that PR! :) Then I will close this one.

Thanks! :)

> I am also working on a perf increase of take_string. That one doesn't have collisions I presume?

No collisions, just go ahead and ping me if you need any help or a review. Note that there may be out-of-bound reads on that code. It would be awesome if those were taken care of :)