@ritchie46 2021-02-03
> Note that there may be out-of-bound reads on that code. It would be awesome if those were taken care of :)

Ah, right so we want to `panic` if we have out-of-bounds access even if `check-bounds == false`. There are quite some cases where I know upfront that all my indexes will be valid. Should we also expose an unsafe variant of the take kernels? Especially in groupby and join operations their performance is paramount.