@jorgecarleitao 2021-02-03
in my experiments leading to #9301 I found no significant perf different between both variants and so did not consider an `unsafe` version, I admit that since the PR already offered a -30%, I also did not bother much ^_^

once all our implementations are `safe`, IMO we should get rid of the take options and potentially offer an `unsafe` version like you said. What imo is not cool is having an implementation whose safety depends on a parameter. :P