@ritchie46 2021-02-03
>  I admit that since the PR already offered a -30%, I also did not bother much ^_^

LOL.. That I can imagine. 

> What imo is not cool is having an implementation whose safety depends on a parameter. :P

Agreed. I must say that I really notice the public API getting more idiomatic as well. It feels good to be able to use iterators for the fastest implementation!